[REFACTOR]: CoupleView의 extension들을 별도 파일로 분리
