+++ POMPOM/POMPOM/UI/CoupleView/CoupleView.swift
@@ -111,210 +111,8 @@ struct CoupleView: View {
     }
 }
 
-// MARK: - Computed properties related to layout
-extension CoupleView {
-    var characterSize: CharacterSize {
-        switch coupleViewModel.sheetMode {
-        case .none:
-            return .large
-        case .mid:
-            return .medium
-        case .high:
-            return .small
-        }
-    }
-    
-    var resetButtonHorizontalPosition: CGFloat {
-        switch coupleViewModel.sheetMode {
-        case .none:
-            return Constant.screenWidth + 44
-        default:
-            return Constant.screenWidth - 31
-        }
-    }
-    
-    var resetButtonVerticalPosition: CGFloat {
-        Constant.screenHeight * (396 / 844) - 90
-    }
-    
-    var characterSpacing: CGFloat {
-        Constant.screenWidth * (33 / 390)
-    }
-    
-    var characterWidth: CGFloat {
-        switch characterSize {
-        case .large:
-            return Constant.screenWidth * (145 / 390)
-        case .medium:
-            return Constant.screenWidth * (114 / 390)
-        case .small:
-            return Constant.screenWidth * (54 / 390)
-        }
-    }
-    
-    var characterOffset: CGFloat {
-        switch characterSize {
-        case .large:
-            return Constant.screenHeight * (93 / 844)
-        case .medium:
-            return Constant.screenHeight * (-29 / 844)
-        case .small:
-            return Constant.screenHeight * (-43 / 844)
-        }
-    }
-    
-    var characterHeight: CGFloat {
-        characterWidth * (215.68 / 114)
-    }
-}
-
-
-//MARK: - Delegates
-extension CoupleView: NetworkDelegate {
-    func showAlertwith(message: String) {
-        coupleViewModel.alertMessage = message
-        coupleViewModel.showAlert.toggle()
-    }
-    
-    func didConnectedPartner() {
-        Task {
-            await partnerClothViewModel.requestPartnerClothes()
-        }
-        coupleViewModel.isConnectedPartner = true
-    }
-}
-
-protocol NetworkDelegate {
-    func showAlertwith(message: String)
-    func didConnectedPartner()
-}
-
 struct CoupleView_Previews: PreviewProvider {
     static var previews: some View {
         CoupleView()
-    }
-}
-
-//MARK: - SubViews
-struct ClothesView: View {
-    @ObservedObject var vm: ClothesViewModel
-    let zIndexIteration: [ClothCategory] = [.hat, .shoes, .bottom, .top]
-    var body: some View {
-        ZStack {
-            ForEach(zIndexIteration) { category in
-                let vm = ClothViewModel(cloth: vm.selectedItems[category], category: category)
-                ClothView(vm: vm)
-            }
-            AccesoriesView(vm: vm)
-        }
-    }
-}
-
-struct AccesoriesView: View {
-    @ObservedObject var vm: ClothesViewModel
-    var body: some View {
-        if vm.isValidItem(with: .accessories) {
-            ZStack {
-                Image(vm.fetchImageString(with: .accessories))
-                    .resizable()
-            }
-            .transition(.slide)
-        }
-    }
-}
-
-// MARK: - Computed properties of subviews
-extension CoupleView {
-    var partnerCharacterView: some View {
-        Button {
-            coupleViewModel.actionSheetPresented = true
-        } label: {
-            ZStack {
-                Image("Gom0")
-                    .resizable()
-                    .frame(width: characterWidth, height: characterHeight)
-                    .opacity(coupleViewModel.isConnectedPartner ? 1 : 0.3)
-                
-                if !coupleViewModel.isConnectedPartner {
-                    Text("초대하기")
-                        .foregroundColor(.orange)
-                } else {
-                    ClothesView(vm: partnerClothViewModel)
-                }
-            }
-            .frame(width: characterWidth, height: characterHeight)
-        }
-        .disabled(coupleViewModel.isConnectedPartner)
-    }
-    
-    var resetButton: some View {
-        Button {
-            myClothViewModel.clearSelectedItem()
-        } label: {
-            Image(systemName: "gobackward")
-                .foregroundColor(Color(UIColor.label))
-                .font(.system(size: 24))
-        }
-        .frame(width: 44, height: 44)
-        .opacity(0.3)
-        .offset(x: resetButtonHorizontalPosition - 0.5 * Constant.screenWidth, y: resetButtonVerticalPosition  - 0.5 * Constant.screenHeight)
-    }
-    
-    var settingsButton: some View {
-        NavigationLink(destination:
-                        SettingsView(showAlert: $coupleViewModel.showAlert, alertMessage: $coupleViewModel.alertMessage, isPartnerConnected: $coupleViewModel.isConnectedPartner)
-        ) {
-            Image(systemName: "gearshape.fill")
-                .foregroundColor(Color(UIColor.label))
-        }
-    }
-    
-    var finishOutfitButton: some View {
-        Button("완료") {
-            myClothViewModel.uploadItem()
-            coupleViewModel.sheetMode = .none
-            
-            Task {
-                await myClothViewModel.requestClothes()
-
-            }
-        }
-    }
-    
-    var cancelOutfitButton: some View {
-        Button("취소") {
-            Task {
-                await myClothViewModel.requestClothes()
-            }
-            coupleViewModel.sheetMode = .none
-        }
-        .foregroundColor(.red)
-    }
-}
-
-// MARK: - methods
-extension CoupleView {
-    private func connectPartner() {
-        codeViewModel.getCode()
-        Task {
-            await myClothViewModel.requestClothes()
-            print("DEBUG: wow")
-        }
-        print(coupleViewModel.isFirstLaunching)
-        Task {
-            await codeViewModel.getPartnerCodeFromServer { partnerCode in
-                print("DEBUG: getPartnerCodeFromServer completion")
-                if partnerCode == "" {
-                    coupleViewModel.isConnectedPartner = false
-                } else {
-                    print(partnerCode)
-                    coupleViewModel.isConnectedPartner = true
-                    Task {
-                        await partnerClothViewModel.requestPartnerClothes()
-                        print("DEBUG: partnerClothViewModel.requestPartnerClothes")
-                    }
-                }
-            }
-        }
     }
 }
